% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/den2Q_dq.R
\name{den2Q_qd}
\alias{den2Q_qd}
\title{convert density function to quantile and quantile density function}
\usage{
den2Q_qd(densityCurves, dSup, t_vec)
}
\arguments{
\item{densityCurves}{n-by-m matrix of density curves}

\item{dSup}{length m vector contains the common support grid of the density curves}

\item{t_vec}{common grid for quantile functions}
}
\description{
convert density function to quantile and quantile density function
}
