% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exp_Map.R
\name{Exp_Map_Barycenter_Method_pw}
\alias{Exp_Map_Barycenter_Method_pw}
\title{Numerical implementation of the pointwise Exponential map}
\usage{
Exp_Map_Barycenter_Method_pw(u, forecast, cdf)
}
\arguments{
\item{u}{The value where the cdf \eqn{\hat{F}} is evaluated}

\item{forecast}{A vector that contains the WAR(p) model forecast result}

\item{cdf}{The quantile grid used in forecasting}
}
\value{
A numeric value \eqn{\hat{F}_t(u)}
}
\description{
This function implements the Exponential map to calculate \eqn{\hat{F}_t(u)} for a fixed \eqn{u}
}
\references{
\cite{Wasserstein Autoregressive Models for Density Time Series, Chao Zhang, Piotr Kokoszka, Alexander Petersen, 2022}
}
\keyword{internal}
