% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalFstat.R, R/short_wass_regress.R
\name{globalFstat}
\alias{globalFstat}
\alias{short_wass_regress}
\title{An internal function used in bootstrap global F test}
\usage{
globalFstat(xfit, Qobs, t_vec)

short_wass_regress(Xfit_df, smoothY)
}
\description{
An internal function used in bootstrap global F test

quick wasserstein regression fitting function with smoothed input
}
\keyword{internal}
