% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.WRI.R
\name{print.summary.WRI}
\alias{print.summary.WRI}
\title{print the summary of WRI object}
\usage{
\method{print}{summary.WRI}(x, ...)
}
\arguments{
\item{x}{a 'summary.WRI' object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print the summary of WRI object
}
