% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrcticQ.R
\name{quadraticQ}
\alias{quadraticQ}
\title{An internal function to do quadratic program in order to make Qfitted nondescreasing}
\usage{
quadraticQ(Qmat, t)
}
\description{
An internal function to do quadratic program in order to make Qfitted nondescreasing
}
\keyword{internal}
