% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quan2den_dq.R
\name{quan2den_qd}
\alias{quan2den_qd}
\title{convert density function to quantile and quantile density function}
\usage{
quan2den_qd(quantileCurves, t_vec)
}
\arguments{
\item{quantileCurves}{n-by-m matrix of quantile curves}

\item{t_vec}{length m vector contains the common support grid of the quantile curves}
}
\description{
convert density function to quantile and quantile density function
}
