% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warEst.R
\name{Sample_ACV}
\alias{Sample_ACV}
\title{Function for calculating sample autocovariance}
\usage{
Sample_ACV(ts, lag)
}
\arguments{
\item{ts}{A numeric vector consisting of sequentially collected data}

\item{lag}{A positive integer indicates the lag of the autocovariance function}
}
\value{
Sample autocovariance
}
\description{
Calculating sample autocovariance of a specified lag for a centered time series
}
\keyword{internal}
