% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warEst.R
\name{WARp_forecast_tangent}
\alias{WARp_forecast_tangent}
\title{Forecast using WAR(p) models}
\usage{
WARp_forecast_tangent(p, acvfs)
}
\arguments{
\item{p}{A positive integer, the order of the WAR(p) model.}

\item{acvfs}{A list that is the output of \code{WARp_acvfs()}}
}
\value{
A numeric vector, the one-step ahead forecast produced by the WAR(p) model in the tangent space.
}
\description{
This is a simplified function that produces the one-step ahead forecasts from WAR(p) models in the tangent space
}
\references{
\cite{Wasserstein Autoregressive Models for Density Time Series, Chao Zhang, Piotr Kokoszka, Alexander Petersen, 2022}
}
\keyword{internal}
