% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warSim.R
\name{warSim}
\alias{warSim}
\title{Generate simulation data}
\usage{
warSim()
}
\value{
A list of:
\item{sample.ts}{one simulation run chosen from \code{sample.ts.full}}
\item{sample.ts.full}{1000 simulation runs, each of which consists of a sample time series (of length 100) of quantile functions generated by a WAR(3) model as specified by the reference paper}
\item{quantile.grid}{the grid over which the quantile functions in sample.ts.full are evaluated}
}
\description{
Generate WAR(p) simulation data sets: samples simulated from a WAR(3) model similar to the specification in Section 4 of the referenced paper.
}
\references{
\cite{Wasserstein Autoregressive Models for Density Time Series, Chao Zhang, Piotr Kokoszka, Alexander Petersen, 2022}
}
